Add basic, incomplete class to handle instances
This class clearly needs a good bit of work, I just threw together a template based around other similar classes. Once we set up GCE with some credits and we get to inspect the responses that these functions are supposed to generate we can make this a whole lot more concrete than it currently is.